Time: 27 ms (87.45%) | Memory: 131.9 MB (97.42%) - LeetSync

+++ 3031-construct-product-matrix/construct-product-matrix.cpp
@@ -3,7 +3,7 @@ public:
     static vector<vector<int>> constructProductMatrix(vector<vector<int>>& grid) {
         const int N=12345, n=grid.size(), m=grid[0].size();
         vector<vector<int>> ans(n, vector<int>(m));
-        uint64_t pref=1, suf=1;
+        long long pref=1, suf=1;
         for(int i=0; i<n; i++)
             for(int j=0; j<m; j++){
                 ans[i][j]=pref;
